+++ src/staticRender.jsx
@@ -1,0 +1,6 @@
+import { renderToStaticMarkup } from "react-dom/server";
+import { StaticPublicPage } from "./app.jsx";
+
+export function renderStaticPublicPage(page) {
+  return renderToStaticMarkup(<StaticPublicPage page={page} />);
+}